refactor(router): refactored a messy directory
